+++ src/Products/PloneMeeting/skins/plonemeeting_styles/plonemeetingpopups.js
@@ -66,7 +66,9 @@ function manageAttendees() {
             closeselector: '[name="form.buttons.cancel"]',
             config: {
                 onBeforeLoad : function (e) {
-                  async_submit_form(this.form);
+                  // odd even for datagridfield
+                  $('table tbody').each(setoddeven);
+                  submitFormHelper(this.form, onsuccess=onsuccessManageAttendees);
                   return true;
                 },
               onClose : function (e) {
@@ -126,49 +128,14 @@ function highlight_attendees(highlight_selector='') {
     'highlight', {}, 2000));
 }
 
-function async_submit_form() {
-  // odd even for datagridfield
-  $('table tbody').each(setoddeven);
-  // onclick for submit button
-  $('input#form-buttons-apply').click(function(event) {
-    event.preventDefault();
-    var data = {'ajax_load': true};
-    $(this.form.elements).each(function(){
-        // use checked instead value for checkbox and radio
-        if (this.type == 'checkbox' && !this.checked) {
-          // unchecked checkboxes are ignored
-          return;
-        }
-        if (this.type == 'radio') {
-          if (this.checked) {
-            data[this.name] = this.value;
-          }
-          else {
-            return;
-          }
-        }
-        // default
-        data[this.name] = this.value;
-      });
-    $.ajax( {
-    type: 'POST',
-    url: this.form.action,
-    data: data,
-    cache: false,
-    async: true,
-    success: function(data) {
-          if (data) {
-            $("div.pb-ajax div")[0].innerHTML = data;
-          }
-          else {
-            refresh_attendees(highlight=null, click_cancel=true);
-          }
-    },
-    error: function(jqXHR, textStatus, errorThrown) {
-      window.location.href = canonical_url();
-    }, } ); });
+function onsuccessManageAttendees(data) {
+  if (data) {
+    $("div.pb-ajax div")[0].innerHTML = data;
+  }
+  else {
+    refresh_attendees(highlight=null, click_cancel=true);
+  }
 }
-
 
 // the content history popup
 function contentHistory() {